Use reactives to render tracks by user input
When the app is loaded, and whenever the user presses the Update button, the track data is rerendered to subset the input data by the user's choices. Much more efficient and useful.

Next up: fix the validator to actually display an error message to the user. It's validating correctly, but I'm not sure why it isn't displaying the message.