Bring preparser up to Python standards.
Been a while since I've used this, but changes made:

* Add requirements.txt to make setting up a venv easier.
* Add Python .gitignore for the preparser subfolder.
* Update preparser script so it isn't a monolith anymore. Use functions
and an actual main() call. Script now passes flask8 and mypy (except for
the NetworkX calls since there's no stubs for it).

I want to update the data to incorporate the Berryessa expansion but I
no longer have access to ArcGIS to edit BART's KML files for what I
need, and it's not worth $100 for this one project. If I find a
cheaper/open-source GIS program I might consider updating it.