Fix double counting in soo data.
BART's original data unfortunately adds some extra hours from the next year in their 2011-2017 data, causing those hours to be double counted.  Remove those.

It's probably best to enforce this in the preparser with
a flag rather than adding checks in the Shiny app to catch this in the future.